Move XLinq specific type description providers into its assembly
Avoids cross-assembly private reflection.
