+++ src/System.Private.Xml.Linq/src/System/Xml/Linq/XAttribute.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.IO;
 
 using CultureInfo = System.Globalization.CultureInfo;
@@ -17,7 +18,7 @@ namespace System.Xml.Linq
     /// An XML attribute is a name/value pair associated with an XML element.
     /// </remarks>
     [SuppressMessage("Microsoft.Naming", "CA1711:IdentifiersShouldNotHaveIncorrectSuffix", Justification = "Reviewed.")]
-    [System.ComponentModel.TypeDescriptionProvider("MS.Internal.Xml.Linq.ComponentModel.XTypeDescriptionProvider`1[[System.Xml.Linq.XAttribute, System.Xml.Linq, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089]],System.ComponentModel.TypeConverter")]
+    [TypeDescriptionProvider(typeof(XTypeDescriptionProvider<XAttribute>))]
     public class XAttribute : XObject
     {
         /// <summary>
