+++ src/System.Private.Xml.Linq/src/System/Xml/Linq/XComponentModel.cs
@@ -2,14 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using System.Xml.Linq;
 
-namespace MS.Internal.Xml.Linq.ComponentModel
+namespace System.Xml.Linq
 {
     internal class XTypeDescriptionProvider<T> : TypeDescriptionProvider
     {
